Add typescript and move some things around
