+++ app/.prettierignore
@@ -1,4 +1,5 @@
 package.json
+tsconfig.json
 node_modules/
 public/*
 build/*
