+++ app/package.json
@@ -5,20 +5,14 @@
   "dependencies": {
     "@material-ui/core": "^4.12.3",
     "@material-ui/icons": "^4.11.2",
-    "@testing-library/jest-dom": "^5.15.0",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
     "ethers": "^5.5.1",
     "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "4.0.3",
-    "web-vitals": "^1.1.2"
+    "react-dom": "^17.0.2"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject",
     "format": "prettier -w ."
   },
   "eslintConfig": {
@@ -40,6 +34,15 @@
     ]
   },
   "devDependencies": {
-    "prettier": "^2.4.1"
+    "@testing-library/jest-dom": "^5.15.0",
+    "@testing-library/react": "^11.2.7",
+    "@testing-library/user-event": "^12.8.3",
+    "@types/jest": "^27.0.3",
+    "@types/node": "^16.11.10",
+    "@types/react-dom": "^17.0.11",
+    "@types/react": "^17.0.36",
+    "prettier": "^2.4.1",
+    "react-scripts": "4.0.3",
+    "typescript": "^4.5.2"
   }
 }
