+++ app/src/App.tsx
@@ -8,8 +8,8 @@ import { ethers } from 'ethers'
 import NavBar from './NavBar'
 import Vote from './Vote'
 import Ranking from './Ranking'
-import DeepStyle from './DeepStyle'
-import { getBackend, CHAIN_RUNNERS_CONTRACT, CHAIN_RUNNERS_ABI, THE23_CONTRACT, THE23_ABI } from './utils'
+import DeepStyle from './components/DeepStyle'
+import { getBackend, CHAIN_RUNNERS_CONTRACT, CHAIN_RUNNERS_ABI, THE23_CONTRACT, THE23_ABI } from './utils/utils'
 
 const theme = createTheme({
   typography: {
