+++ app/src/components/NavBar.tsx
