+++ app/src/components/Vote.tsx
